MAILBOX-362 Group all non event DTOs to DTOs scala file
